Make the h2 database properties configurable
Hitherto, AAA has had the limitation of static h2 database configuration.
Since the H2Config is stored in a Java file, the only way to modify h2
properties is to recompile the bundle with the new properties specified,
which is obviously not too user friendly.  Note, the credentials here are
not ODL AAA credentials;  rather, they are the credentials for the backing
h2 data store.  In most cases, they may be divergent from the credentials you
use to access ODL's restful web services!

This change introduces a new concept called h2.cfg, which is located in
the etc folder.  The idea behind h2.cfg is to expose some of the h2 configuration
so the user can modify as appropriate to his/her setup.  h2.cfg was designed
as a simple key/value list, and utilizes java.util.Properties.loadProperties(...)
to load and parse property values.  This approach was preferred over the
ManagedService approach, which was tried initially, but found to be inadequate
since the configuration ManagedService did not necessarily load before the h2
service causing the defaults to be used.  This level of non-determinism made
the java.util.Properties approach preferred and the obvious choice.

This change allows configuration of the following database properties:

1) username  (kept the same by default for backwards compatibility)
2) password  (kept the same by default for backwards compatibility)
3) connectionString  (kept the same by default for backwards compatability)

The first two values are trivial to set.  The third value ("connectionString")
is a bit trickier;  the recommendation is to keep this value in the form:

connectionString = jdbc:h2:[[INSERT_YOUR_IDMLIGHT_DB_PATH_HERE]]

for example:

connectionString = jdbc:h2:./aaa.db

The first two parts should not be altered;  if you wish to use a different JDBC
implementation, then you should consider pursuing configuration based on
org.opendaylight.aaa.shiro.realm.ODLJdbcRealm instead.

An important note is that if these values are changed, then you need to make the
corresponding changes to your database (i.e., add the user, change the password,
etc.).  If you wish to not use defaults, it is recommended to either:

1) lay down the h2.cfg file as part of an external installer prior to installing
   odl-restconf or odl-aaa-shiro.

or

2) After changing the values in h2.cfg, stop the karaf process, delete the database
file (idmlight.db.mv.db by default), then start the karaf process again.  The
h2 initialization code will automatically rebuild a default version of the database.

This adds a necessary level of security abstraction on top of the existing
implementation;  hard-coded credentials are never good, even if they are local
to a file-based database.

Change-Id: I158b45066b40f9e06d0b132e4946c4a55fe767f0
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>