Update patch set 1
Patch Set 1:

At some point we need to re-organize the aaa project and remove the unused and not maintained source code. Keep having unused and unmaintained code will affect Opendaylight deployment "it could raise a serious security issues in future". We should start work on that for Carbon release.

Patch-set: 1