Update patch set 1
Patch Set 1:

> Maybe it worth reading the comments that occurred in that patch:
 > https://git.opendaylight.org/gerrit/#/c/38480/ regarding not using
 > aaa-parent. I believe AAA didn't want to go there in the past,
 > maybe it has change..

In the past we were resistant to changing the parent since having a local parent was useful for local project dependency management.  The main motivation behind not moving was that it was easier to add/change versions of dependencies in a local parent than trying to get the change in upstream odlparent (The classic "States' Rights" argument).  This was largely driven by a once relatively inactive odlparent committer team;  this has obviously changed since.  I believe we should revisit this topic (flip flopping now due to more upstream support).  Adding Mohamed for his feedback, as he is also part of the active AAA committer team.

Patch-set: 1