Update patch set 1
Patch Set 1:

As I could remember, the problem with moving to config-parent. We had a dependency conflict happen between the one of aaa bundles and the config-parent dependancy that makes the aaa release job fail at jenkins during the unit test. 
The idea for aaa-parent is we have different modules using their own implementations  with different dependancies aaa-parent will be easier to handle any kind of dependancy conflict could happen as before. However, as we will clean up the aaa for this release and might remove some bundles. we can move to config-parent later on.

Patch-set: 1