Bug 6956 - Do not wrap Guava as a bundle in the feature definition
Guava should only be provided as a dependency, and don't need to be
provided within the feature definition as a bundle.
Doing so could potentially have bad effect: e.g. DLUX feature pulls in Guava
the same way, which will trigger the Guava bundle to be refresh, thus
the AAA bundles tied to the feature pulling it in will be refreshed as
well, and this is corrupting functionalities as per as the reported
BUG.

Change-Id: If519c51c4a47a5b7e9e76f793ee81bba565d0d16
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>