Update patch set 1
Patch Set 1:

Guys you'll probably want to carefully review and ideally even locally test this.. if I screwed up something here, it will obviously break everything. -- FYI I'm not doing this for fun, but will be using this in an upcoming future Gerrit which will be dependant on this.

Patch-set: 1
Label: Verified=0