De-static-ify H2Store's IdmLightConfig and intro. proper design
This is required to be able to configure a H2Store with an
IdmLightConfig, e.g. from the upcoming new CLI tool.

The intention then is to use this to subsequently introduce a real JDBC
Connection Pool on top of this new API.  As a first step, the changes
introduced here (should, hopefully) functionally still make it behave
exactly as the current implementation.

Change-Id: Ia28f5eb9c154c5c74fcef7ad285eee8b6be32ffb
Signed-off-by: Michael Vorburger <vorburger@redhat.com>