Update patch set 8
Patch Set 8:

> > What about granting usernames specific permissions based on role?
 > > Don't we need the capability to be able to grant a username
 > 'admin'
 > > level role/privileges?
 > 
 > OK, I'll add that when I pick this up - after the 2 other are
 > merged..

Not to mention that just creating a user does nothing.  You will not be able to access anything with just creating the user;  you need to grant the user a role on a specific domain (in this case SDN).  Thus, you may want to add business logic to automatically grant the created user the "admin" or "user" role on "sdn".   Good catch, @Tim.

Patch-set: 8