Update patch set 8
Patch Set 8:

> Hi,
 > So basically this jar required to have a physical access to the
 > h2.db which mean it will run at the same ODL host. is this right ?
 > just to make sure I understand correctly.
 > If so, why not adding these functionalities to karaf CLI instead ?
 > you can still add, update and delete users, roles ..etc without
 > knowing the user password. And this can be used for all datastore
 > types not only h2.
 > Just think about it Michael, this will really help. Also I can help
 > you on that if  you need help

aaa-cli relies on the fact that the user will input the existing password prior to inputting the new password.  That is not idempotent, similar to how idmlight.py is not idempotent.  The point of this is to basically add an administrative hook that can be run to gen some predefined users/passwords in an idempotent way.

Patch-set: 8