Update patch set 2
Patch Set 2: Code-Review-1

Hi,
aaa has different data stores, there are mdsal, H2Store and Cassandra. all of them implement IIDMStore and run it as the auth service for ODL. why do you made this changes instead of just refer to IIDMStore and get the service ? this is kind of duplicated work.

Patch-set: 2