Deprecate the authz model
The authz model causes confusion in the community;  several people want to use
authorization functionality but it was never implemented correctly.  The original
contributor has since stopped participating in upstream AAA.  The model does not
work correctly.

The AAA team does want to add authz other than shiro based, but this will have to
be handled by severely changing the existing model, such that keeping it around is
quite silly since it never worked properly in the first place.

Change-Id: Id3035ca71ec483e8d8c887179a635439898b7e64
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>