Addition of a new http oriented authorization model
The existing aaa-authz feature was not functional, and the model was deficient.  This is the
introduction of a new authorization model which can better support end users' needs.  This
drop is for the API only;  the implementation will be done separately once agreement on the
API is reached.

The implementation details have not yet been decided;  there is a very likely chance that we
will utilize Shiro's url based authorization for this use case.  The existing BrokerFacade
work is tricky and can be made to work, but is not very reliable at present.  The BrokerFacade
is also oriented towards DOM API and not HTTP/S operations, which we are more interested in
enforcing with this work.

Change-Id: I889f394d59f343d6a65fd5e6a85d78f22a712fe1
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>