cherry-pick Migrate aaa-shiro to utilize archetype setup
This change addresses the fact that the archetype was not used to
create the aaa-shiro module.  This is due to the fact that it was
much heavier weight than what was needed at the time.  However,
utilization of the archetype does allow for many advantages including:

1) Explicit separation of api and impl into two separate locations.
   Since prior to this change mostly Apache Shiro abstractions were used
   instead of homegrown ones, this wasn't particularly useful.  However,
   with aaa-shiro growing, this is becoming increasingly necessary.

2) Dependence on config-parent for bundles.  This gets us a lot for
   free, including genericizising on best practices.

3) The possibility to create aaa-shiro features/karaf/IT/commands more
   easily.  For now, this patch comments out the features & karaf
   section as they are pretty much duplicates of the existing top-level
   odl-aaa-shiro feature which is already defined.  In the future, it
   would be nice to enumerate some of the archetype-oriented features:
   - ui
   - rest
   - api

This change is mostly cleaning up and preparing for expansion of the
aaa-shiro bundle.  Existing functionality was stuck in the aaa-shiro module,
but the package names were not updated to utilize impl.  Likewise, the module
is called "aaa-shiro" and not "aaa-shiro-impl" since other projects already depend
on the former name.  The package names were not updated to utilize impl
as the names are used in configuration of the module itself, and
changing them would cause forwards/backwards compatibility issues.  In the future,
we may want to move them but provide existing classes that extend the impl ones.

This change is used with a few others to add in dynamic authorization.  Because
the change is adding in a significant security enhancement, it is a good
candidate to cherry pick back into stable/boron.

Change-Id: I16f1efed8b83e764362ae6d19b0a69d1b1c6cbec
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>