Remove RBAC rule implementation
RBAC rules were a concept that existed to automatically restrict some endpoints.  However,
they were not terrible useful and not mutable.  The point was to just restrict a
certain subset of endpoints for security purposes (i.e., the IdM endpoints).

This change removes unused concepts and makes a few minor fixes to existing code:
* make local vars final when appropriate
* better logging
* utilize File.separator instead of hardcoding "/"

Change-Id: If3a1f100e8a2b265be71cfb4722b64c76aacad34
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>