Remove the broken aaa-authz module and corresponding models
aaa-authz no longer works.  The service was originally designed as a DataBrokerFacade
to aid in MD-SAL based authorization.  This should be considered again in the future,
but for now we are removing this functionality since it is broken and could cause
confusion.  Since the mechanism currently doesn't work, we feel it is better to
completely remove the feature rather than deprecate something that doesn't do anything.

To get HTTP based authorization, Shiro integration is recommended.  This is explained
in etc/shiro.ini, as well as upstream in Shiro documentation.

Change-Id: I6c822860e9b28a9577eb88c87c7c10120e4c9a97
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>