Remove cassandra based store
This is dead code, since there appears to be no way to configure cassandra as
the default backing data store.  Thus, the code exists, but is never instantiated.

Even the feature "odl-aaa-cassandra*" does not properly load the cassandra based
backing data store.

Since the feature doesn't appear to work or add anything extra, it is being removed.
We are providing an alternative dropin backed by mdsal instead, which should
be used instead.

Change-Id: I30c5231753544b170fb05bb461734cbc34efec8a
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>