Update README surrounding accounting
Accounting has been greatly improved since the Beryllium release with
the AuthenticationListener implementation.  This updates the README
for accounting only.  Further updates will be submitted surrounding
Authentication and Authorization prior to release.

This patch is a canddate for master, Boron and Beryllium, and will
add valuable information for users of each of these releases.

Change-Id: I6f7bc6ce6a4d178eb7c00810102795d8c8b9c987
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>