Update patch set 5
Patch Set 5: Code-Review-1

There is no need to add the CLI and feature bundles. There is already a aaa-cli and feature bundles we can move them instead. It's good to have the same archetype setup but wait till we clean up the unused bundles then we can move bundles to fit with the archetype setup. Moreover if you are not planning to add api to aaa-shiro/api/ bundles there is no need for it.

Patch-set: 5
Label: Code-Review=-1