Update patch set 1
Patch Set 1:

> > Please don't touch features while the Karaf 4 migration is in
 > > progress.
 > 
 > Is this just in general or wide reaching changes?  I have a patch
 > that will touch features.xml, so I want to make sure it won't break
 > the world :).  This is the patch:
 > 
 > https://git.opendaylight.org/gerrit/#/c/50694/
 > 
 > Not merged yet since I am still working on more Unit Testing.

We did say we didn't want the Karaf 4 migration to stop people from making progress on other features, so no, it's not a general rule. Basically any change to features needs to be propagated to the Karaf 4 patches, so it just means more book-keeping; for a small change such as this one (from AAA's perspective anyway) it seems to me we can afford to wait.

Patch-set: 1