Update patch set 7
Patch Set 7:

> |otherwise we end up with a failure loading odl-aaa-cli since the
 > bundle has a
 > |strong dependency on Karaf 3.
 > Can you elaborate more on that, what is the error you see when you
 > updated the karaf4.vsersion at aaa-cli bundle.  I'm not sure how
 > can I help you

At this point the only help I'm asking for on the AAA patch is some reviews, especially regarding the aaa-shiro and aaa-idmlight changes (see the commit message).

As far as aaa-cli goes, remember that the intent of the Karaf 4 migration is to continue providing Karaf 3 features in parallel for the time being. This means that we can't simply change ${karaf.version} to ${karaf4.version}: the resulting odl-aaa-cli feature would then be incompatible with Karaf 3. We need to either arrange things so that odl-aaa-cli (with aaa-cli) works in both Karaf 3 and 4, or keep the old odl-aaa-cli with aaa-cli for Karaf 3, and add a new odl-aaa-cli for Karaf 4 (with a duplicated aaa-cli4 bundle, using ${karaf4.version}. I've chosen the latter approach because it's more straightforward, but the former might be possible with an import override in the maven-bundle-plugin configuration in aaa-cli.

Trying to load the existing odl-aaa-cli in Karaf 4 produces

missing requirement [org.opendaylight.aaa.cli/0.5.0.SNAPSHOT] osgi.wiring.package; filter:="(&(osgi.wiring.package=org.apache.karaf.shell.commands)(version>=3.0.0)(!(version>=4.0.0)))"

Patch-set: 7