Remove felix.dependencymanager from aaa-h2-store
This makes life easier for my friends working hard on the Karaf 4 mig.

The switch of Import-Package from explicit list to exclusion list is
indirectly related; I may also have been possible to leave that as is
and just remove "org.apache.felix.*" from that list, but flipping this
around like this seems much better, to me.

Change-Id: I0ee3d3765db9d4ec466939b4862a38d957cdb451
Signed-off-by: Michael Vorburger <vorburger@redhat.com>