Shiro authz RBAC roles read from keystone.
The user roles within a domain are extracted from the auth token, which
is granted by keystone. Local authorization then takes place, depending
on shiro's configuration.

The authentication request to Keystone was changed, turning the auth
mechanism into 'scoped' authentication. The scope used is to the domain
in which the user is authenticating.

By using this scoped authentication, Keystone replies in the token with
the list of roles of that particular user in the specified domain.

Change-Id: I511aa2f49c40fc7b1ac7ce7abd495f38a68d7581
Signed-off-by: Miguel Duarte <miguel.duarte.de.mora.barroso@ericsson.com>