Update patch set 3
Patch Set 3:

So in regards to your las comment Mohamed, I want to make sure it was made in relation to patch set 3 and not to original patch set 4 (which upload was unintended and made the automatic verification fail).

Also I want to note that in patch set 3 I am adding a dependency to odl-aaa-cert feature in the odl-aaa-shiro feature itself and on those features that load the aaa-shiro bundle, only.

I tried your suggested approach of <Embed-Dependency>. With that I have managed to just add the dependency to odl-aaa-cert in odl-aaa-shiro feature. This is what I used in aaa-shiro pom:

            <Embed-Dependency>aaa-cert|bcprov-jdk15on|aaa-encrypt-service;scope=compile</Embed-Dependency>
            <Embed-Transitive>true</Embed-Transitive>

bcprov-jdk15on and aaa-encrypt-service are dependencies of aaa-cert. Instead of transitive handling, I guess we could add the corresponding Embed-Dependency on aaa-cert bundle also.

Let me know if this is ok with you. Honestly, I dont quite understand the preference of doing this versus adding the appropriate <bundle> directives on the appropriate features.

Thanks for your time.

Patch-set: 3