KeystoneAuthRealm: reuse http client
Jersey documentation advises against building multiple clients of the
same characteristics and advertises that is perfectly safe to use the
same client over multiple threads.

Moreover, each time the SSL context is obtained, it may synchronize the
keystores from the filesystem or the database. KeystoneAuthRealm
assumes a dynamic certificate handling.

The purpose of this commit is to reuse the http client to some degree
while refreshing it rather frequently due to the dynamic certificate
handling.

Change-Id: Ifa188ad6f2a99e64c5d560d58057428aab37df26
Signed-off-by: Jaime Caamaño Ruiz <jaime.caamano.ruiz@ericsson.com>