Bug 8214: Expose Service Interface for AAA
Hitherto, the tie between AAA and its downstream consumers has been
nebulous.  The instantiation of javax servlets requires runtime
dependencies to be resolved.  Bug 8214 exposes a race condition
in which RESTCONF attempts to resolve AAAFilter, but the AAA
service is not yet instantiated.  To resolve this, the solution
is to:

1) expose a AAAService interface.  In this case, that is quite
easy since we already have the necessary implementation,
AAAFilter.  Thus, all that is done is the extraction of the
service interface.

2) Advertise the service to the OSGI registry via blueprint.
This is a quick 3 liner.

A corresponding change will be submitted to the downstream
consumer (RESTCONF) to depend on AAAService.

Change-Id: I5d8208ddbe817963ce596ecc34ad38ff0f760410
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>