Clean read-only transactions
Read-only transactions are documented to be a resource and should
be closed as soon as they are not needed. Failing to do so
triggers warnings in tell-based protocol in CDS, as they trigger
inefficient GC-triggered cleanup.

Change-Id: Ia8468704a75fd42c105f3fba49f817d5c2e0ec30
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 3a2b698c58ab6cee9f75e84f307298cf75ab5553)