Bug 8214:  Avoid throwing NPE when DataBroker is null
MDSALDynamicAuthorizationFilter is instantiated by shiro when
the web container is brought up.  shiro has no knowledge of
the DataBroker, so the dependency cannot be injected in shiro.ini.

shiro.ini needs to die, and is tracked by Bug 7793.  For now,
to avoid throwing the NPE (in the case that the DataBroker hasn't
been brought up yet), just fail-close until the dataBroker is
made available (i.e., AAAShiroProvider.newInstance(DataBroker)
is called).

Change-Id: I28e2eb0780f4f7761a5e2f81d167afc5a4b10cba
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>