Update patch set 1
Patch Set 1:

The other option is to not inject this filter in shiro.ini, and add it once AAAShiroProvider is instantiated.  I don't like that approach since it is less clear from an external standpoint what is going on, but it may be a way to get us over this hump.

Patch-set: 1