Bug 8379: Remove deprecated TokenAuthFilter
TokenAuthFilter was deprecated in either Beryllium or Boron.
It has been kept around for compatibility reasons, but it
is no longer needed.  This patch removes the Filter class,
moves the inner UnauthorizedException to its own class
(better anyway), and removes some dead tests that were
testing the old filter functionality.  It is clear that
the tests were testing the old Filter mechanism instead
of AAAShiroFilter, since AUTH_FILTERS was set to
TokenAuthFilter.class.getName().  Thus, the test is no
longer appropriate.

Change-Id: I08295daccc13bd9ac9113a8cf55e779ca1001775
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>