sssd functionality removal
odl-authn-sssd was deprecated in Boron.  The feature was removed
in Carbon, but the underlying bundles were kept just in case
(i.e., one could install manually but it wasn't a present
feature).  This completes the removal since adequate time has
been given for consumers to move on to the new method, Shiro.

Change-Id: Idb95c0f41a3db0300fd8fbd7272ec3ab842f9c45
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>