MDSAL based AuthorizingRealm API only
Adds an MDSALRealm API only.  None of the impl changes are
in this patch.

The MDSALRealm will be a clustered data store implementation
for IDM.  The old store MD-SAL based store will be removed
in future patches, but will need to be done carefully since
it appears (from mailing lists) that people somehow still
use it.  We will need to publish some sort of transition
plan for people.

Change-Id: I84aebf4c172dbaf63ad8f24c5b4d380a1e7baa3b
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>