Bug 8437: move / eliminate aaa-authn-basic
aaa-authn-basic contained functionality for decrypting authentication
details from a basic authentication request.  However, it really was
overkill for such a small amount of functionality.  This patch
moves the needed code to the aaa-shiro bundle.

This is part of a broader effort to reduce the number of bundles in
AAA, and to make sure all bundles are activated using blueprint.
In this case, we eliminate a bundle and make sure the aaa-authn-basic
functionality is now loaded via blueprint since aaa-shiro services
are already activated by blueprint.

Change-Id: Ie96aff9a99ea6428e3c31483027af46c3433fbbf
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>