Bug 8437: Move oauth2 functionality into the shiro module
odl-aaa-shiro is the only consumer of the oauth2 functionality in ODL.
The fact that aaa-authn-sts is aggregated as a separate bundle for
this functionality is confusing and superfluous.  This change moves
the functionality, does some good renaming, and puts in some docs
surrounding the current status of aaa-authn-sts and aaa-shiro.

This change was tested manually with odl-restconf and works
correctly.

In the future, in a follow up, aaa-authn-sts will be completely
removed.  For now it is still keeping track of the various "services"
that we are working to converge.

Change-Id: Id9cc5bd70e0866d72b712ddc559c1cd0154268d9
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>