DO NOT MERGE: Collapse Bundles
One of the major focuses of the Nitrogen release surrounds service stability
in Karaf.  Hitherto, AAA has made use of a wide variety of service wiring
techniques, some effective, some not so great.  The service wiring really
was composed of three forms:
1) blueprint (good)
2) CSS (removed in previous patches and replaced by #1)
3) felix dependency management, which is really old and required long waits
   to properly wire various components

Since #2 was already removed, this patch removes #3, which is the wiring of
services through ServiceTrackerCustomizers and the singleton ServiceLocator.
ServiceLocator still exists in this patch;  to remove it here would make the
patch much too large, but it will be refactored in a followup patch.

For now, what is done is the following:
1) removal of aaa-authn bundle.  This bundle contained Builders and utility
classes common to aaa-idmlight and aaa-mdsal-store bundles.  They are
implementation details, and thus not candidates for odl-aaa-api.  They were
put in odl-aaa-shiro, as eventually odl-aaa-mdsal-cluster will eventually be
removed and replaced with functionality built directly into odl-aaa-shiro
(see https://git.opendaylight.org/gerrit/#/c/56744/)

2) removal of aaa-authn-basic.  This bundle really didn't provide much
more than HttpAuthBasic, which is really the only common impl of TokenAuth
used in AAA.  This bundle's functionality was relocated to odl-aaa-shiro.

3) removal of aaa-authn-sts.  STS stands for "single token store".  This
bundle also provided limited functionality, and really didn't make sense as
a completely separate service.

This patch focuses on the consolidation of several aaa bundles and services
into something more manageable directly in aaa-shiro.

Future work will:
1) remove the existing mdsal store in favor of the MDSAL based
   AuthorizingRealm.  Basically, the old mdsal store utilizes a lot of hacky
   code, reflection, etc. which can be cleaned and redefined to be more
   managable.  This will be removed in Oxygen, as we forgot to deprecate
   it in Nitrogen.  Messaging will be delivered to ensure that people use
   the new Realm approach once it is available.  This is again, not going
   to be the default store anyway in Nitrogen.

2) further refactoring of Services so that the aaa-idmlight and aaa-h2-store
   bundles are included directly in aaa-shiro.  Basically, it is really
   against OSGi architecture to have this many services providing such
   limited functionality.  Instead, we ought to provide them as part of a
   single service, and make configuration through clustered-app-config,
   cfg files, or some other more sane way than service injection.  The
   fact remains that we really want a more statically wired configuration
   that is consistent and easy to understand.

This is just part of many patches to clean the existing AAA implementation.

Change-Id: I740a667278f2ff64daff5e3602ddc5586fe23733
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>