Add functionality to change jolokia password
Jolokia authentication is still unfortunately limited to
authMode=basic.  In the meantime, functionality is added
to allow easy change of this password through the exposed
jolokia config file org.jolokia.osgi.cfg.

In the future this password will be tied to AAA directly.
For now, just expose an easy utility to change it.

Change-Id: I53b7f67326dcc6dbd09a6e42b0f055ff8d1766d8
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>