Update patch set 1
Patch Set 1:

I am suggesting we deprecate this one and use the one i pushed in my other patch.  This one has concepts that were never expressed (isEnabled etc), and the impl is horrid.  We could also continue to utilize this API but swap out the backing impl, but that is going to be a lot of work for not much gain.  I am open to suggestions though...

Patch-set: 1