Update patch set 8
Patch Set 8:

> > (3 comments)
 > >
 > > just few comments, overall looks fine.
 > 
 > Yes, I'd like to see how can configure the KeystoneAuthRealm and
 > parameters needed to setup this realm.

Hi David,

Yes, I am adding some more default config to the xml file.  I also will have a utility that is meant to convert your old shiro.ini to the XML representation, which can than be put to the configuration data store... it is in this patch (not updated version since it is a WIP):

https://git.opendaylight.org/gerrit/#/c/59588/

Hope this alleviates any fears :).

Patch-set: 8
Label: Verified=0