Update patch set 19
Patch Set 19:

> Patch Set 19: Code-Review+1

This is failing due to a broken downstream.  That downstream (of-config IIRC) is a dependency for another project and thus cannot easily be removed from autorelease.  There are patches to fix it but no one actually looking after it...

Going to recheck one last time, and if it is still occurring will look at how to either rip it out or fix it in some way.

Patch-set: 19
Label: Verified=0