Update patch set 6
Patch Set 6:

My point behind the last comment is I didn't quite understand that this was just for tokens.  It looked like it was also for Users at first, but I was wrong about that.  Can we just update the yang to make that more clear?  This makes more sense now that you explained it.

Patch-set: 6