Bug 8717 Remove the existing non-functional mdsal impl
This implementation was designed by someone who has long since
left the project and has not been maintained since Lithium release
minus minor fixes to service activation and AAA API changes.

It was deprecated last release with:
https://git.opendaylight.org/gerrit/#/c/59189/

This removes the implementation so it cannot be used.  We
plan on adding a revamped version of the MDSAL store in
Oxygen timeframe.

Change-Id: I705924aac88ea317e1378fbeee4117f52d5e3904
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>