Update patch set 9
Patch Set 9:

> well, it's okay let's discuss on this week meeting. and again if u
 > have idea in mind just push a patch to express ur idea.

Well thats not how it works;  you aren't supposed to expect me to retroactively push a patch after this is merged to express my opinions.  You are supposed to wait to merge it until my comments have been satisfied.  Would've been glad to hold off merging other patches to avoid merge conflicts in the meantime.  The fact is that we are trying to move away from these older constructs  (defined in current aaa-authn-api), and this just ties us to them longer.  Not what we discussed in our last call!

Patch-set: 9