Update patch set 2
Patch Set 2:

This is just an initial ugly draft I revived.  Basically, in 500 lines I have a basic MdsalRealm which does not depend on all the old constructs/APIs in aaa-authn-api.  This will allow us to begin deprecating all that old stuff, and to transition to this eventually.  It will also allow us to eventually remove the REST endpoints located at "/auth".  This is why I was so against the other patch;  it essentially made us depend on all the old things that I think we should trash sooner rather than later.

Patch-set: 2