AAA-149: migrate ServiceProxy to the correct package name
This migrates ServiceProxy to the correct package.  AAA-149 is
being done in several small steps in order to ensure that the
changes do not affect downstream consumers.  This change was
tested against NETCONF to ensure there was no downstream
breakages (which makes sense since NETCONF does not directly
manipulate ServiceProxy and actually only affects it through
AAA code).

Change-Id: If142f4fe1ddc91cf844fb8d0ca3a6496f0828efd
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>