AAA-151:  Invalidate claim cache for CLI initiated changes
The AAA IDM REST handlers already invoke the clearing of the
IdmLightProxy claimCache upon user update, but the CLI Handler did
not handle this properly.  This change manually invokes the claim
invalidation.

Access to AAA CLI commands requires access to the karaf console via
client or SSH.  Thus, system level access is needed to some degree
to invoke such commands.  The major threat involves the action of
an administrator changing a user password via CLI, and then still
being able to invoke the old password because the claimCache is not
purged.  This is why a reboot of karaf solves this issue;  the
claimCache is flushed.  We should not expect our users/administrators
to assume this in the future;  this change explicitly invokes
the claimCache flush.

AAA-151 affects Carbon, Nitrogen, and master.

Change-Id: Ie3101d4dc2bbb6d0db91ddad4d567a41d041150d
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>