Switch to using gson for JSON serialization
This patch changes JSON provider to GSON, skipping jackson
packaging completely.  GSON is used by many upstream ODL
projects and is the desired single JSON library for the
future.

Some unit tests surrounding Handlers needed to be ignored
temporarily due to classpath issues;  some of the overlapping
technology is tested in integration-test so the AAA team
feels it is okay to temporarily disable for now.

For now, this change exposes a provider package for the
GsonProvider so that other projects running into the same
issue can utilize it without copying and pasting code.
Later, it will be moved somewhere more intelligent.

Change-Id: I6033980d0fdaa31be32e2e77a0b9f869a755246b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>