AAA-155: don't force non-null input on currentUser cache lookup
On the first CLI command issued no auth users are cached in either the
SessionsManager cache nor AaaCliAbstractCommand's static authUser. We
must take input from the user and do the lookup in the identity store.
Subsequent commands will use the now cached user until the entry is
evicted. getCurrentUser should handle the null input and be explicit
about the nullability of it's return value.

Change-Id: I18291e25723f428d2e27f79184d957f7715357f8
Signed-off-by: Evan Zeller <evanrzeller@gmail.com>