Temp. Remove MDSAL from aaa-app-config
Per AAA meeting discussions, it is probably better to delegate
Token Processing to the MdsalRealm or a corresponding pre-filter.
This has been done in the past for things like OAuth2 Token
processing.  This change purely removes the option for MDSAL
backed store from the aaa-app-config, since the existing impl is
provided separately through MdsalRealm.  Right now, choosing
MDSAL as an option actually only instantiates skeleton code, so it
is better to leave this option out in this release for
security purposes.

Change-Id: Ia32809f02865af8f96e0bdacbd20d064055114fe
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>