Update patch set 7
Patch Set 7:

> Because AAA is not working.  It is explained in the related email.

got any links explaining the blocker, or perhaps even a JIRA?  This change was flagged internally (Sam & JamO) as a possible concern for destabilizing the Oxygen release.. and we're not all following all emails, sorry!  From our vantage point as consumer in netvirt CSIT, AAA seems to work just fine, so we appreciate further clarification re. criticality, and/or assurances why this is truly low risk and can still be changed this late in the game.

> If NETCONF needs jackson, it needs to bring it in on its own, not depend on AAA bringing it in.

sounds great!  But even NETCONF does seem to working for us now; finally.

Patch-set: 7