AAA-159: Switch to using gson for JSON serialization
Due to jackson incompatibility issues with other web env
elements (namely jax-rs), this change switches the serdes
for AAA endpoints to gson.  The motivation is two fold:

1) fix the immediate issue with the rest endpoints
2) align on a common JSON serdes framework in ODL

Since yangtools and others already utilize gson, and gson
seems to be a lot more friendly from a provider stand-
point, this change is the best solution to the given bug.

This patch does not completely remove jackson, since RESTCONF
depends on us bringing it in.  This will be another multi-
step process:

1) this patch
2) convert restconf to GSON or just add the jackson deps
   there (features/odl-aaa-shiro/pom.xml changes in last
   patch)
3) remove jackson dependencies from AAA

Change-Id: Id969ab11282513fc314b98cd2a3487327250113f
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>