AAA-159: Switch to using gson for JSON serialization
Due to jackson incompatibility issues with other web env
elements (namely jax-rs), this change switches the serdes
for AAA endpoints to gson.  The motivation is two fold:

1) fix the immediate issue with the rest endpoints
2) align on a common JSON serdes framework in ODL

Since yangtools and others already utilize gson, and gson
seems to be a lot more friendly from a provider stand-
point, this change is the best solution to the given bug.

Change-Id: Id969ab11282513fc314b98cd2a3487327250113f
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>